Revert 412ecef..8ce1c6e
This rolls back to commit 412ecef8b95fcc5d375d3e0e7ac8f85a85dba5c2.

+++ src/com/PP/IntelliSeg/Util/SelectionFunction.java
@@ -17,8 +17,8 @@ public class SelectionFunction {
 	public static double score(String gram, Set<Integer> startSet, String str_rep)
 	{		
 		//compute parts
-		double L = gram.length();
-		double F = startSet.size();
+		double L = (double) gram.length();
+		double F = (double) startSet.size();
 		double T = F*L;
 		boolean[] used = new boolean[str_rep.length()];
 		for(int start : startSet) {
